Code/Django/projects/pjt07_instaclone/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.7
django-appconf==1.0.4
django-bootstrap-v5==1.0.0
django-imagekit==4.0.2
importlib-metadata==2.1.1
pilkit==2.0
Pillow==8.1.2
pytz==2021.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
zipp==3.4.1